Refactor: Moved settings and helper functions to their own files
