+++ core/tests/APITest.php
@@ -2,7 +2,7 @@
 
 
 require_once(__DIR__."/../../vendor/autoload.php");
-require_once(__DIR__."/../const.php");
+require_once(__DIR__."/../../settings.php");
 require_once(__DIR__."/../api.php");
