+++ core/url.php
@@ -1,7 +1,7 @@
 <?php declare(strict_types=1); namespace Neosluger;
 
 
-require_once(__DIR__."/const.php");
+require_once(__DIR__."/../settings.php");
 
 
 /** @class URL
