+++ web/main.php
@@ -3,8 +3,9 @@
 
 ini_set("display_errors", '1');
 require_once($_SERVER['DOCUMENT_ROOT']."/vendor/autoload.php");
+require_once($_SERVER['DOCUMENT_ROOT']."/core/helper-functions.php");
 require_once($_SERVER['DOCUMENT_ROOT']."/core/url.php");
-require_once($_SERVER['DOCUMENT_ROOT']."/core/const.php");
+require_once($_SERVER['DOCUMENT_ROOT']."/settings.php");
 
 
 /** @fn find_page (string $uri): string
