+++ web/pages/api.php
@@ -3,9 +3,9 @@
 
 ini_set("display_errors", '1');
 require_once($_SERVER['DOCUMENT_ROOT']."/vendor/autoload.php");
-require_once($_SERVER['DOCUMENT_ROOT']."/core/const.php");
 require_once($_SERVER['DOCUMENT_ROOT']."/core/api.php");
 require_once($_SERVER['DOCUMENT_ROOT']."/core/url.php");
+require_once($_SERVER['DOCUMENT_ROOT']."/settings.php");
 
 
 function process_api_petition (): void
