+++ web/pages/index.php
@@ -3,7 +3,7 @@
 
 ini_set("display_errors", '1');
 require_once($_SERVER['DOCUMENT_ROOT']."/vendor/autoload.php");
-require_once($_SERVER['DOCUMENT_ROOT']."/core/const.php");
+require_once($_SERVER['DOCUMENT_ROOT']."/settings.php");
 
 
 function render (): void
