+++ web/pages/inspect-result.php
@@ -2,7 +2,7 @@
 
 
 ini_set("display_errors", '1');
-require_once($_SERVER['DOCUMENT_ROOT']."/core/const.php");
+require_once($_SERVER['DOCUMENT_ROOT']."/settings.php");
 
 
 function filter_short_link_handle (string $query): string
